refactor feed related packages
