+++ app/src/androidTest/java/com/group80/uoftinder/LogoutActivityUITest.java
@@ -20,7 +20,7 @@ import org.junit.runner.RunWith;
 
 @RunWith(AndroidJUnit4.class)
 @LargeTest
-public class LogoutTest {
+public class LogoutActivityUITest {
     @Rule
     public ActivityScenarioRule<LoginActivity> activityRule =
             new ActivityScenarioRule<>(LoginActivity.class);
