+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.UofTinder"
         tools:targetApi="31">
         <activity
-            android:name=".ProfilePicUploadActivity"
+            android:name=".view_layer.ProfilePicUploadActivity"
             android:exported="true">
             <intent-filter>
                 <category android:name="android.intent.category.DEFAULT" />
@@ -24,7 +24,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".feed.AcademicFilterActivity"
+            android:name=".view_layer.AcademicFilterActivity"
             android:exported="true" >
             <intent-filter>
                 <category android:name="android.intent.category.DEFAULT" />
@@ -61,7 +61,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".feed.NoNewRecommendation"
+            android:name=".view_layer.NoNewRecommendation"
             android:exported="true" >
 
             <intent-filter>
@@ -69,7 +69,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".feed.RecommendationView"
+            android:name=".view_layer.RecommendationView"
             android:exported="true">
             <intent-filter>
                 <category android:name="android.intent.category.DEFAULT" />
