+++ app/src/main/java/com/group80/uoftinder/create_account_use_case/CreateAccountController.java
@@ -4,7 +4,7 @@ import android.widget.EditText;
 
 import com.google.android.material.chip.Chip;
 import com.google.android.material.chip.ChipGroup;
-import com.group80.uoftinder.entities.User;
+import com.group80.uoftinder.entities_layer.User;
 
 import java.util.LinkedList;
 import java.util.List;
