+++ app/src/main/java/com/group80/uoftinder/entities_layer/User.java
@@ -1,4 +1,4 @@
-package com.group80.uoftinder.entities;
+package com.group80.uoftinder.entities_layer;
 
 import android.net.Uri;
