+++ app/src/main/res/layout/activity_profile_upload.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfilePicUploadActivity">
+    tools:context=".view_layer.ProfilePicUploadActivity">
 
     <TextView
         android:id="@+id/profileUploadActivityTitleTextView"
